% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust.R
\name{clust.all.both}
\alias{clust.all.both}
\title{clust.all.both}
\usage{
clust.all.both(x, y, rMax, period, expected, observed, Time, utm, byrow)
}
\arguments{
\item{x}{x coordinates (easting/latitude); if utm coordinates, scale to km.}

\item{y}{y coordinates (northing/longitude); if utm coordinates, scale to km.}

\item{rMax}{set max radius (in km)}

\item{period}{vector of periods or years in dataset. Should be imported as a factor.}

\item{expected}{vector of expected counts. Expected counts must match up with the year and observed vectors.}

\item{observed}{vector of observed counts. Observed counts must match up with the year and expected vectors.}

\item{Time}{Number of time periods or years in your dataset. Must be declared as numeric.}

\item{utm}{default is TRUE. If FALSE, then will run long/lat data}

\item{byrow}{default is True. If data should be imported by column then set to FALSE}
}
\value{

}
\description{
This function runs both the space and space-time Lasso model. This function is to be run on observed data. A separate function (clust) is the helper function which will have 
flexibility to specify the space or spacetime or both models to be run (TODO).
}

