% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustDiagnostics.R
\name{clust.diagnostics}
\alias{clust.diagnostics}
\title{Suite of function for cluster detection diagnostics - incluster diagnostics}
\usage{
clust.diagnostics(incluster, threshold, nullmod, ...)
}
\arguments{
\item{incluster}{inherited object from detect.incluster.ic}

\item{threshold}{detection threshold set by user}

\item{nullmod}{default is NULL. If not null, then null model results will be returned}
}
\value{
list of detection
}
\description{
clust.diagnostics
}
\examples{
set <- detect_set(lassoresult.qp.s, vectors.sim, rr, Time, x, y, rMax, center, radius)
incluster.qp.q <- detect.incluster(lassoresult.qp.s, vectors.sim, rr, set,timeperiod, Time, nsim, x, y, rMax, center, radius, IC = "ic")
detect.qp.s <- list(clust.diagnostics(incluster.qp.s, threshold[1]),clust.diagnostics(incluster.qp.s, threshold[2])
}

