% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_sim.R
\name{clust.sim.all.both}
\alias{clust.sim.all.both}
\title{clust.sim.all.both}
\usage{
clust.sim.all.both(x, y, rMax, period, expected, observed, Time, nsim, center,
  radius, risk.ratio, timeperiod, utm, byrow, threshold, theta = theta,
  nullmod = nullmod, ...)
}
\arguments{
\item{x}{x coordinates (easting/latitude); if utm coordinates, scale to km.}

\item{y}{y coordinates (northing/longitude); if utm coordinates, scale to km.}

\item{rMax}{set max radius (in km)}

\item{period}{vector of periods or years in dataset. Should be imported as a factor.}

\item{expected}{vector of expected counts. Expected counts must match up with the year and observed vectors.}

\item{observed}{vector of observed counts. Observed counts must match up with the year and expected vectors.}

\item{Time}{Number of time periods or years in your dataset. Must be declared as numeric.}

\item{nsim}{Number of simulations you would like to run}

\item{center}{can be a single center or for multiple clusters, concatenate them. Max three TODO extend this}

\item{radius}{radius for the cluster you want in the simulation}

\item{risk.ratio}{setting for what the risk ratio should be in the cluster to be detected by the simulation}

\item{timeperiod}{time period where the cluster  starts in the simulation
we will only be looking at periods 2 and 5. If multi_period is TRUE, then we will instead consider timeperiod through period_end (timeperiod:period_end). Following the same example,
this would mean we look at periods 2, 3, 4, and 5.}

\item{theta}{default is 1000. Can add in overdispersion to simulated model by changing this value.}

\item{nullmod}{if TRUE, then null models will be run. Otherwise, default is null.}

\item{space}{space and space-time. Default is to run all four models: quasi-poisson and poisson for both space and space-time. User can specify, space = space,
space = spacetime, or space = both.}
}
\value{
returns list of lists
}
\description{
This function runs both the space and space-time Lasso model simulations for all 4 models simulataneously: Quasi-Poisson vs. Poisson in both space and space-time.
This function is to be run on simulated data and all four models are run on the same simulated set. 
A separate function (clust.sim) can be used for running simulations on individual models and (clust) can be used for observed data.
}

