% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustDiagnostics.R
\name{detect.incluster}
\alias{detect.incluster}
\title{detect.incluster}
\usage{
detect.incluster(lassoresult, vectors.sim, rr, set, timeperiod, Time, nsim, x,
  y, rMax, center, radius, IC = c("aic", "aicc", "bic", "ic"),
  under = FALSE, nullmod)
}
\arguments{
\item{lassoresult}{List of QBIC, QAIC, QAICc estimates from the mylasso.sim function}

\item{vectors.sim}{dataframe of initial vectors of the observed and expected counts that went into simulation function}

\item{rr}{risk ratio matrix that was used in the simulation}

\item{set}{result of detect_set function}

\item{timeperiod}{span of timeperiods in simulation}

\item{Time}{total number of time periods}

\item{nsim}{number of simulations}

\item{x}{x coordinates}

\item{y}{y coordinates}

\item{rMax}{max radius}

\item{center}{center of centroid for cluster}

\item{radius}{radius of the cluster}

\item{IC}{the information criteria you would like to be returned. Options are: IC = aic or IC = qaic; IC = aicc or IC = qaicc; IC = bic or IC = qbic; IC = ic or IC = qic 
(for aic/aicc/bic and qaic/qaicc/qbic, respectively).}

\item{under}{default is FALSE. If risk.ratio is less than one (under-risk)}

\item{nullmod}{default is NULL. If not null, then null model results will be estimated and returned.}
}
\value{
returns
}
\description{
This function will calculate the percent of simulations which correctly identify elements in cluster based on (Q)AIC, (Q)AICc, and (Q)BIC. The user can specify
if they want to only return one of these criterion or all three for further analysis.
}

