% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_set.R
\name{detect_set}
\alias{detect_set}
\title{Set initial conditions to enable detection diagnostic functions}
\usage{
detect_set(lassoresult, vectors.sim, rr, Time, x, y, rMax, center, radius,
  nullmod, ...)
}
\arguments{
\item{lassoresult}{List of QBIC, QAIC, QAICc estimates from the mylasso.sim function}

\item{vectors.sim}{dataframe of initial vectors of the observed and expected counts that went into simulation function}

\item{rr}{risk ratio matrix that was used in the simulation}

\item{x}{x coordinates}

\item{y}{y coordinates}

\item{rMax}{set maximum radius for a cluster}

\item{center}{center of the simulation cluster}

\item{radius}{radius of the simulated cluster}

\item{nullmod}{indicator for whether or not the simulation being run is the null model or not. Default is not null.}
}
\value{
returns a list of 1) indx_truth, which is a vector of indices where our cluster
should be, 2) indx, which contains the index of cluster as determined where the risk ratio in the rr matrix is not 1,
3) rr.simBIC, rr.simAIC, rr.simAICc - the risk ratios as determined by BIC, AIC, and AICc (respectively), and 4) alphaBIC, alphaAIC, alphaAICc - the calculated
background risk rates as determined by BIC, AIC, and AICc or QBIC, QAIC, QAICc.
}
\description{
This function will extract and set the necessary vectors in order to detect what percent of clusters
in the simulation were detected (true positives), what percent of clusters were incorrectly in the background (false negatives),
and what percent were incorrectly identified as clusters (false positives).
}

