% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpoisson.R
\name{dpoisson}
\alias{dpoisson}
\title{Poisson Distribution Function}
\usage{
dpoisson(y, lambda, E0)
}
\arguments{
\item{y}{observed values}

\item{lambda}{vector of exp(each column of each potential path). Assumes you exponentiate prior to using this function}

\item{E0}{expected counts (the offset). If in simulation, then this should be the E0 counts for each simulation. These E0 have been standardized such that E0 = E*(sum(y)/sum(E)) with the scale.sim()/scale() function depending on if a simulation is being run or not}
}
\value{
returns a a vector (or a list of vectors) of the Poisson log-likelihood for each proposed path of the Lasso tuning parameter values
}
\description{
This is the main distriution function for our model. This assumes we have a Poisson fixed effect and Gamma random effect. In order to deal with constraints from the Lasso function, we use the Poisson distirbution function here and account for overdispersion in the QIC.
}
\examples{
set.seed(1)
E <- rnegbin(n = 20,mu = 15,theta = 1000)
theta = 1000
y <- rnegbin(E, theta=theta)
E0 <- E*(sum(y)/sum(E))
lambda <- exp(c(rep(0,15),rep(0.4,5)))
dpoisson(y, lambda, E0)
}

