\name{entropy.of.partition}
\alias{entropy.of.partition}
\title{Entropy of a partition}
\usage{
entropy.of.partition(partition)
}
\arguments{
  \item{partition}{: input partition. Any labelling scheme
  will do, hopefully.}
}
\value{
Shannon entropy for the input partition. Not foolproof if
partition is empty.
}
\description{
Entropy of a partition
}
\author{
M.GELGON
}

