\name{generate.partition.for.each.feature}
\alias{generate.partition.for.each.feature}
\title{Generate partitions of a given data set for each Feature}
\usage{
generate.partition.for.each.feature(dataset, number.partitions, number.clusters)
}
\description{
Generate partitions of a given data set for each Feature
}

