\name{generate.partitions.many.kmeans}
\alias{generate.partitions.many.kmeans}
\title{Generate partitions of a given data set with many kmeans}
\usage{
generate.partitions.many.kmeans(dataset, number.partitions, number.clusters)
}
\description{
Generate partitions of a given data set with many kmeans
}

