% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setRR.R
\name{get_rr}
\alias{get_rr}
\title{get_rr}
\usage{
get_rr(lassoresult, vectors, init, E1, Time, sim = TRUE, ...)
}
\arguments{
\item{lassoresult}{List of QBIC, QAIC, QAICc estimates from the mylasso function}

\item{init}{initial settings}

\item{E1}{initial expected counts}

\item{Time}{number of time period}

\item{sim}{default is TRUE. Signals whether these values are to be extracted from a simulated model or real data.}

\item{vectors.sim}{dataframe of initial vectors of the observed and expected counts}
}
\value{
This returns a list of the risk ratios (observed over expected) as determined by 1) pure observed/expected counts,
2) observed based on QBIC path/expected; 3) observed based on QAIC path/expected; 4) observed based on QAICc path/expected.
}
\description{
This function will extract the relative risk ratios as determined by observed counts, QBIC, QAIC, and QAICc, respectively. This method determines the 
relative risk by averageing the relative risk in each simulation over the number of simulations. This is in contrast to $setRR$ where the expected counts
are first average over the number of simulations and then we determine the relative risk to the initial expected counts. Both methods return the same results,
however this method returns a smoother background ratio that does not vary across time periods as much
}

