\name{mutual.information.two.partitions}
\alias{mutual.information.two.partitions}
\title{Mutual information between two partitions}
\usage{
mutual.information.two.partitions(partition1, partition2)
}
\arguments{
  \item{partition1}{a partition}

  \item{partition2}{a partition}
}
\value{
the mutual information between the input partitions
}
\description{
Mutual information between two partitions
}
\author{
M.GELGON
}

