\name{normalized.mutual.information}
\alias{normalized.mutual.information}
\title{Normalized Mutual information between two partitions
A similarity between partitions that
takes values between 0 (independent) and 1 (identical)}
\usage{
normalized.mutual.information(partition1, partition2)
}
\arguments{
  \item{partition1}{a partition}

  \item{partition2}{a partition}
}
\value{
the normalized mutual information
}
\description{
Normalized Mutual information between two partitions A
similarity between partitions that takes values between 0
(independent) and 1 (identical)
}
\author{
M.GELGON
}

