\name{normalizedMutualInformation}
\alias{normalizedMutualInformation}
\title{Normalized Mutual information between two partitions
A similarity between partitions that
takes values between 0 (independent) and 1 (identical)}
\usage{
normalizedMutualInformation(partition1, partition2)
}
\description{
Normalized Mutual information between two partitions A
similarity between partitions that takes values between 0
(independent) and 1 (identical)
}

