% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overdisp.R
\name{overdisp}
\alias{overdisp}
\title{Calculate overdispersion for quasi-likelihood models}
\usage{
overdisp(object)
}
\arguments{
\item{object}{object of class glm}
}
\value{
returns sparse matrix of 1's
}
\description{
This function calculates the overdispersion parameter for the QIC 'c' overdispersion parameter.
}
\examples{
offset <- glm(Yx ~ 1 + as.factor(vectors$Period) + offset(log(Ex)), family=poisson)
overdisp.est <- overdisp(offset_reg)
}

