% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyplot.R
\name{plotmap_S}
\alias{plotmap_S}
\title{#Space-only plotting}
\usage{
plotmap_S(pdfname, res, obs, sub, ...)
}
\arguments{
\item{pdfname}{pdfname of what the output pdf will be called}

\item{res}{resultant list from clust_ function}

\item{obs}{if observed is to be plotted or oracle from simulation}

\item{sub}{optional parameter if you want to just use the function for plotting different vectors}
}
\description{
#Space-only plotting
}

