% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probmap.R
\name{probmap}
\alias{probmap}
\title{Mapping colors to create a black/white probability map}
\usage{
probmap(lassoresult, vectors.sim, rr, nsim, Time, colormap = FALSE, ...)
}
\arguments{
\item{lassoresult}{List of QBIC, QAIC, QAICc estimates from the mylasso.sim function}

\item{vectors.sim}{dataframe of initial vectors of the observed and expected counts that went into mylasso.sim function}

\item{rr}{risk ratio matrix that was used in the simulation}

\item{nsim}{number of simulations}

\item{Time}{number of time period}

\item{colormap}{default is FALSE. Signals whether the probabilities should directly be mapped to the red-blue color scheme. If this is false,
only the probability values will be returned. If true, then the probability values and the mapped colors will be returned in a list.}
}
\value{
returns vector which calculated the number of time the cluster was correctly identified out of the simulations
}
\description{
This function will create a probability map based on simulation data. In each simulation, it identifies where a cluster was selected,
compared to the background rate. It then average over the number of simulations, giving us a matrix which ranges from 0 to 1 in probability.
To map this probabilities into a color scheme, please see the $colormapping$ function and select probmap=TRUE. TODO integrate all of this
into a workflow and extend to observed data, not only simulated data.
}

