% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setVectors.R
\name{setVectors}
\alias{setVectors}
\title{Creates a List Arranged by Time Period with Expected and Observed Counts and Time Period}
\usage{
setVectors(period, expect, observed, Time, byrow = TRUE)
}
\arguments{
\item{period}{vector of periods or years in dataset. Should be imported as a factor.}

\item{expect}{vector of expected counts. Expected counts must match up with the year and observed vectors.}

\item{observed}{vector of observed counts. Observed counts must match up with the year and expected vectors.}

\item{Time}{Number of time periods or years in your dataset. Must be declared as numeric.}

\item{byrow}{default is set to TRUE. Data from the dataset should be imported by row. This is most often the case
when you have a dataframe ordered by an identifier and then the period/time frame within that id listed chronologically (in panel format by identifier).
If you are simulating data and have each observed/expected vector separate and create the period vector with repetitions of each time
period by group, this should be set to false.}
}
\value{
This function returns a list of expected and observed counts along with the period.
}
\description{
Creates a List Arranged by Time Period with Expected and Observed Counts and Time Period
}
\examples{
period1 <- c(rep("1",5),rep("2",5))
period2 <- rep(seq(1,2),5)
expected <- rnegbin(n = 10,mu = 15,theta = 1000)
observed <- rnegbin(expected, theta=1000)
Time = 2
setVectors(period1, expected, observed, Time, byrow=TRUE))
setVectors(period2, expected, observed, Time, byrow=FALSE)
}

