\name{variation.of.information}
\alias{variation.of.information}
\title{variationOfInformation
A similarity between partitions that has some forgotten advantage}
\usage{
variation.of.information(partition1, partition2)
}
\arguments{
  \item{partition1}{a first partition}

  \item{partition2}{a first partition}
}
\value{
the variation of information between partitions
}
\description{
variationOfInformation A similarity between partitions that
has some forgotten advantage
}
\author{
M.GELGON
}

