% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_sim.R
\name{vectors.space.sim}
\alias{vectors.space.sim}
\title{Simulation functions for space and space-time cluster detection with lasso.}
\usage{
vectors.space.sim(x, Ex, YSIM, Time, init, ...)
}
\arguments{
\item{x}{vector coordinates (unique regardless of time period)}

\item{Ex}{list of simulated and standardized expected counts}

\item{YSIM}{simulated observed}

\item{Time}{number of time periods}

\item{init}{initial list of vectors, inherited from function setVectors.}

\item{timeperiod}{explicit vector of timeperiods}
}
\value{
returns space-time
}
\description{
vectors.space.sim
}
\details{
This function will collapse a space-time vector onto space only
}

