% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust.R
\name{vectors_space}
\alias{vectors_space}
\title{vectors_space}
\usage{
vectors_space(x, Ex, Yx, Time, init)
}
\arguments{
\item{x}{vector coordinates (unique regardless of time period)}

\item{Ex}{list of simulated and standardized expected counts}

\item{Yx}{observed}

\item{Time}{number of time periods}

\item{init}{initial list of vectors, inherited from function setVectors.}
}
\value{
returns space-time
}
\description{
This function will collapse a space-time vector onto space only
}
\details{
Detect a cluster in space or spacetime using Lasso
}

