\name{dataPairWiseSimilarityMatrix}
\alias{dataPairWiseSimilarityMatrix}
\title{Compute the pair-wise similarity matrix between all lines of a partition matrix}
\usage{
dataPairWiseSimilarityMatrix(partitions)
}
\arguments{
  \item{partitions}{two partitions, what format ?}
}
\value{
a similarity matrix, no clue
}
\description{
Compute the pair-wise similarity matrix between all lines
of a partition matrix
}
\author{
M.GELGON
}

