\name{generate.partition.for.each.feature}
\alias{generate.partition.for.each.feature}
\title{Generate partitions of a given data set for each Feature}
\usage{
generate.partition.for.each.feature(dataset, number.clusters)
}
\arguments{
  \item{dataset}{a data frame containing a dataset}

  \item{number.clusters}{how many clusters}
}
\value{
a set of partitions
}
\description{
Generate partitions of a given data set for each Feature
}

