\name{generate.partitions.many.kmeans}
\alias{generate.partitions.many.kmeans}
\title{Generate partitions of a given data set with many kmeans, using several random initializations.}
\usage{
generate.partitions.many.kmeans(dataset, number.partitions, number.clusters)
}
\arguments{
  \item{dataset}{a data frame containing a dataset}

  \item{number.partitions}{how many partitions}

  \item{number.clusters}{how many clusters}
}
\value{
a set of partitions
}
\description{
Generate partitions of a given data set with many kmeans,
using several random initializations.
}

