% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overdisp.R
\name{overdisp}
\alias{overdisp}
\title{Calculate overdispersion for quasi-likelihood models}
\usage{
overdisp(offset_reg, sim, floor)
}
\arguments{
\item{offset_reg}{object of class glm}

\item{sim}{for a simulation run this takes on TRUE, else it should be FALSE. Default is set to TRUE.}

\item{floor}{Default is set to TRUE so model will not estimate any under-dispersion where phi < 1. The floor is set for phi to be 1. In a case where this}
}
\value{
returns phi overdispersion parameter
}
\description{
This function calculates the overdispersion parameter for the QIC 'c' overdispersion parameter.
}
\examples{
offset <- glm(Yx ~ 1 + as.factor(vectors$Period) + offset(log(Ex)), family=poisson)
overdisp.est <- overdisp(offset_reg)
}

